Create shared ResetFormButton component and update dependent paths
