+++ Client/src/Views/Question/DefaultQuestionForm.scss
@@ -27,11 +27,6 @@
     padding-top: 1em;
   }
 
-  &ResetFormButton.btn {
-    font-size: 1.1em;
-    margin: 1em 0;
-  }
-
   &ShowHideGroup {
     border: 1px solid #eaeaea;
     background-color: #f0f0f0;
